requirements.txt
fastmcp
starlette
uvicorn[standard]
openpyxl
numpy
requests
python-dotenv
pytest
